Move Every() from submission package to new schedule package
It could be more widely used than just in the submission package.
